+++ system-parachains/asset-hubs/asset-hub-polkadot/src/treasury.rs
@@ -29,8 +29,11 @@ parameter_types! {
 	pub TreasuryAccount: AccountId = Treasury::account_id();
 }
 
-pub type TreasuryPaymaster =
-	crate::pay::LocalPay<NativeAndAssets, TreasuryAccount, xcm_config::LocationToAccountId>;
+pub type TreasuryPaymaster = system_parachains_common::pay::LocalPay<
+	NativeAndAssets,
+	TreasuryAccount,
+	xcm_config::LocationToAccountId,
+>;
 
 impl pallet_treasury::Config for Runtime {
 	type PalletId = TreasuryPalletId;
@@ -51,8 +54,9 @@ impl pallet_treasury::Config for Runtime {
 	type BalanceConverter = AssetRateWithNative;
 	type PayoutPeriod = PayoutSpendPeriod;
 	#[cfg(feature = "runtime-benchmarks")]
-	type BenchmarkHelper =
-		crate::pay::benchmarks::LocalPayArguments<xcm_config::TrustBackedAssetsPalletIndex>;
+	type BenchmarkHelper = system_parachains_common::pay::benchmarks::LocalPayArguments<
+		xcm_config::TrustBackedAssetsPalletIndex,
+	>;
 	type BlockNumberProvider = RelaychainDataProvider<Runtime>;
 }