+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add `cumulus-pallet-weight-reclaim` to all Kusama system parachains for storage weight reclaim functionality ([#941](https://github.com/polkadot-fellows/runtimes/pull/941))
-- Add local pay module implementation for Asset Hub Polkadot treasury with XCM V3 to V5 conversion support for asset transactions
+- Add local pay module implementation for Asset Hub Polkadot treasury with XCM V3 to V5 conversion support for asset transactions ([#1019](https://github.com/polkadot-fellows/runtimes/pull/1019))
 
 ## [2.0.3] 21.11.2025